requirements.txt
Django>=4.2.4,<4.3.0
djangorestframework>=3.14.0,<3.15.0
psycopg2>=2.9.7,<2.10.0
django-cors-headers>=4.2.0,<4.3.0
